+++ CHANGES.rst
@@ -36,6 +36,12 @@ Bug fixes:
   This solves a problem, where Editors couldn't see inactive content, even though they had the required permission on a subpath of the portal (e.g. a subsite).
   [thet]
 
+- Test: Wrong use of assertTrue in testResourceRegistries.
+  [jensens]
+
+- Fix issue popped iup after fix of use of assertTrue in testResourceRegistries: insert-before in legacy resource import was broken.
+  [jensens]
+
 
 5.1b2 (2017-02-20)
 ------------------